backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
supabase==2.4.5
python-dotenv==1.0.1
pydantic==2.6.4
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
pyyaml==6.0.1
stripe==8.8.0
